Refactor. Todo model to separate package
